requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.3
cachetools==6.2.4
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
Django==5.2.8
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
exceptiongroup==1.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
groq==0.36.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
inflection==0.5.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
